Revise static result example
Using `aria-describedby` on elements like `<div>` or `<p>` is not broadly supported in screen readers.

VoiceOver and NVDA do not recognize the connection when a non-form field, button, link or landmark use `aria-describeby` to point to another element.

JAWS does reveal the connection, but requires a user to press <kbd>JAWS key + ALT + R</kbd> to read the descriptive text. (JAWS auto-reads descriptive text when used on one of the mentioned supported elements).

As the example isn’t well supported, and isn’t truly necessary if a  document is setup appropriately, the revised pattern changes the movie name text to be within an `<h3>`, adds a line of dummy info text, and places the rating to be within a `<p>`, with additional “Rated:” text to provide context for why someone would be hearing “3 stars”